# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AccessScope": {
        "additionalProperties": false,
        "properties": {
          "empresa_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Id"
          },
          "fundo_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fundo Id"
          },
          "lote_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lote Id"
          },
          "modulo_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modulo Id"
          }
        },
        "title": "AccessScope",
        "type": "object"
      },
...
      "ApiError": {
        "description": "Error legible para la app y para Swagger.",
        "example": {
          "detail": "Descripción del error"
        },
        "properties": {
          "detail": {
            "description": "Motivo por el que la operación no pudo completarse",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "AuthUser": {
        "properties": {
          "alcances": {
            "items": {
              "$ref": "#/components/schemas/AccessScope"
            },
            "title": "Alcances",
            "type": "array"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "permisos": {
            "items": {
              "type": "string"
            },
            "title": "Permisos",
            "type": "array"
          },
          "rol": {
            "title": "Rol",
            "type": "string"
          },
          "usuario_id": {
            "title": "Usuario Id",
            "type": "integer"
          }
        },
        "required": [
          "usuario_id",
          "email",
          "nombre",
          "rol"
        ],
        "title": "AuthUser",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "TokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "TokenPair",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API interna para registrar evaluaciones de app-campo en PostgreSQL. Incluye autenticación administrativa y lectura paginada para el panel Angular.",
    "title": "Aqu Anqa · API de campo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "iniciarSesionAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email o contraseña incorrectos"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Credenciales inválidas o sesión vencida."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar usuarios"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Base no disponible."
          }
        },
        "summary": "Inicia sesión administrativa",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "obtenerSesionAdmin",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email o contraseña incorrectos"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Credenciales inválidas o sesión vencida."
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar usuarios"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Base no disponible."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Devuelve la identidad activa",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "renovarSesionAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email o contraseña incorrectos"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Credenciales inválidas o sesión vencida."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "No se pudo consultar usuarios"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Base no disponible."
          }
        },
        "summary": "Renueva una sesión administrativa",
        "tags": [
          "Autenticación"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Servidor actual (local)",
      "url": "/"
    }
  ],
...
}